pre-last warning fixes for 0.9.8

+++ shedskin/lib/builtin/tuple.hpp
@@ -40,8 +40,8 @@ template<class T> tuple2<T, T>::tuple2(tuple2<T, T> *p) {
 template<class T> tuple2<T, T>::tuple2(str *s) {
     this->__class__ = cl_tuple;
     this->units.resize(len(s));
-    int sz = s->unit.size();
-    for(int i=0; i<sz; i++)
+    size_t sz = s->unit.size();
+    for(size_t i=0; i<sz; i++)
         this->units[i] = __char_cache[((unsigned char)(s->unit[i]))];
 }
